# contracts/openapi/traktor.yaml
openapi: 3.1.0
info:
  title: Traktor API
  version: 0.1.0
  description: >
    Единый контракт REST API Traktor (source of truth). Из него генерируются
    Dart-клиент (packages/api_client), серверные интерфейсы (Go) и документация.
    Меняется контракт → регенерируются обе стороны. Ломающие изменения ловит CI.
servers:
  - url: https://api.traktor.am/v1
    description: prod (домен — после выбора Тиграном)
  - url: https://api.stage.traktor.am/v1
    description: stage
...
paths:
...
  /categories:
    get:
      tags: [catalog]
      summary: Дерево категорий (работы и техника)
      description: >
        Открыто без входа: справочник нужен и гостю в режиме «просто посмотреть».
        Названия приходят сразу на трёх языках — переключение языка в приложении
        не требует нового запроса.
      operationId: listCategories
      parameters:
        - name: kind
          in: query
          schema: { type: string, enum: [work, unit] }
          description: work — виды работ (визард задания), unit — техника (визард техники)
        - name: flat
          in: query
          schema: { type: string, enum: ["1"] }
          description: "1 — плоский список вместо дерева (для поиска по названию)"
      responses:
        "200":
          description: Категории
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Category" }
        "400": { $ref: "#/components/responses/Problem" }
  /categories/{id}:
    get:
      tags: [catalog]
      summary: Одна категория
      operationId: getCategory
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Категория
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "404": { $ref: "#/components/responses/Problem" }
...
  /equipment/my:
    get:
      tags: [catalog]
      summary: Моя техника (ТЗ §2.5)
      operationId: myEquipment
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Список машин исполнителя, свежие сверху
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Equipment" }
  /equipment/drafts:
    post:
      tags: [catalog]
      summary: Начать визард техники
      operationId: createEquipmentDraft
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                categoryId: { type: string, format: uuid }
      responses:
        "201":
          description: Черновик создан
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Equipment" }
  /equipment/{id}:
    get:
      tags: [catalog]
      summary: Карточка техники
      operationId: getEquipment
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Карточка
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Equipment" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
    patch:
      tags: [catalog]
      summary: Шаг визарда или правка карточки
      description: >
        Приходят только изменённые поля. Пока карточка на проверке, правки
        отклоняются: модерация смотрит на то, что было отправлено.
      operationId: patchEquipment
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Equipment" }
      responses:
        "200":
          description: Сохранено
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Equipment" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /equipment/{id}/submit:
    post:
      tags: [catalog]
      summary: Опубликовать карточку техники
      description: >
        С документами карточка уходит на проверку (≤24 ч), без них публикуется
        сразу как «Без проверки»: работать можно, но в скоринге ниже.
      operationId: submitEquipment
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Опубликовано или отправлено на проверку
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Equipment" }
        "422":
          description: Карточка заполнена не полностью
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationProblem" }
  /equipment/{id}/archive:
    post:
      tags: [catalog]
      summary: Снять технику с площадки
      operationId: archiveEquipment
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Снято
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Equipment" }
  /media/uploads:
    post:
      tags: [media]
      summary: Временные ссылки на загрузку файлов (ADR-5)
      description: >
        Клиент грузит файл прямо в хранилище по временной ссылке, минуя наш
        бэкенд: фотографии весят мегабайты. Принимаем JPEG, PNG, WebP, HEIC и
        PDF; имя файла придумывает сервер.
      operationId: createUploadLinks
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contentType]
              properties:
                contentType: { type: string, example: image/jpeg }
                folder: { type: string, enum: [equipment, jobs, docs, misc] }
                count: { type: integer, minimum: 1, maximum: 8, default: 1 }
      responses:
        "200":
          description: Ссылки на загрузку
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/UploadLink" }
        "415": { $ref: "#/components/responses/Problem" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string }
      description: Гарантия «двойной тап ≠ две операции» (ТЗ §4.3).
  responses:
    Problem:
      description: Ошибка (RFC 9457 problem+json) с человекочитаемым detail для тоста
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    Equipment:
      type: object
      required: [id, status]
      properties:
        id: { type: string, format: uuid }
        categoryId: { type: string, format: uuid }
        categoryName: { $ref: "#/components/schemas/LocalizedName" }
        brand: { type: string, example: JCB }
        model: { type: string, example: 3CX }
        year: { type: integer, minimum: 1980 }
        specs:
          type: object
          additionalProperties: true
          description: "значения по specTemplate категории"
        priceHour: { type: integer, format: int64 }
        priceShift: { type: integer, format: int64 }
        priceDay: { type: integer, format: int64 }
        minHours: { type: integer }
        delivery: { type: integer, format: int64 }
        crewSize: { type: integer, description: "бригада рабочих с техникой" }
        crewPrice: { type: integer, format: int64 }
        photos:
          type: array
          items: { type: string }
          maxItems: 8
        status:
          type: string
          enum: [draft, pending, verified, unverified, rejected, archived]
        rejectReason: { type: string }
        draftStep: { type: integer, minimum: 1, maximum: 4 }
        wins: { type: integer, description: "выиграно работ этой машиной" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    UploadLink:
      type: object
      properties:
        key: { type: string }
        uploadUrl: { type: string, description: "временная ссылка для PUT" }
        publicUrl: { type: string, description: "постоянный адрес файла" }
        expiresIn: { type: integer, example: 900 }
...
    Category:
      type: object
      required: [id, kind, slug, name, icon, specTemplate]
      properties:
        id: { type: string, format: uuid }
        parentId: { type: string, format: uuid, nullable: true }
        kind: { type: string, enum: [work, unit] }
        slug: { type: string, example: work-earth }
        name: { $ref: "#/components/schemas/LocalizedName" }
        icon: { type: string, description: "Имя иконки Phosphor из design_system (эмодзи запрещены)" }
        specTemplate:
          type: array
          items: { $ref: "#/components/schemas/SpecField" }
        sortOrder: { type: integer }
        children:
          type: array
          items: { $ref: "#/components/schemas/Category" }
    LocalizedName:
      type: object
      required: [hy, ru, en]
      properties:
        hy: { type: string }
        ru: { type: string }
        en: { type: string }
    SpecField:
      type: object
      required: [key, type]
      description: Поле характеристик — из него клиент строит поле формы.
      properties:
        key: { type: string, example: bucket }
        type: { type: string, enum: [number, text, select, bool] }
        unit: { type: string, example: "м³" }
        min: { type: number }
        max: { type: number }
        options:
          type: array
          items: { type: string }
        label_hy: { type: string }
        label_ru: { type: string }
        label_en: { type: string }
        required: { type: boolean }
...
    ValidationProblem:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            fields:
              type: object
              additionalProperties: { type: string }
              description: "поле → что именно не так; экран подсвечивает их"
...
    Problem:
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string, description: "человекочитаемо, годится в тост" }